refactor: Rename app from LifeRhythm to DailyRhythm
- Update package name from life_rhythm to daily_rhythm
- Change application ID to com.dailyrhythm.dailyrhythm across all platforms
- Update app display name in all platform configurations
- Update documentation and README
- Reorganize Android package structure

Platform updates:
- Android: Updated manifest, build files, and package name
- iOS: Updated Info.plist and bundle identifier
- macOS: Updated AppInfo.xcconfig
- Linux: Updated CMakeLists.txt and GTK window title
- Windows: Updated CMakeLists.txt, Runner.rc, and window creation
- Web: Updated manifest.json and index.html
