+++ android/app/build.gradle.kts
@@ -13,7 +13,7 @@ if (keystorePropertiesFile.exists()) {
 }
 
 android {
-    namespace = "com.liferhythm.liferhythm"
+    namespace = "com.dailyrhythm.dailyrhythm"
     compileSdk = flutter.compileSdkVersion
     ndkVersion = flutter.ndkVersion
 
@@ -27,7 +27,7 @@ android {
     }
 
     defaultConfig {
-        applicationId = "com.liferhythm.liferhythm"
+        applicationId = "com.dailyrhythm.dailyrhythm"
         minSdk = flutter.minSdkVersion
         targetSdk = flutter.targetSdkVersion
         versionCode = flutter.versionCode
