+++ android/app/src/main/kotlin/com/dailyrhythm/dailyrhythm/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.liferhythm.liferhythm
+package com.dailyrhythm.dailyrhythm
 
 import io.flutter.embedding.android.FlutterActivity
