+++ linux/runner/my_application.cc
@@ -46,11 +46,11 @@ static void my_application_activate(GApplication* application) {
   if (use_header_bar) {
     GtkHeaderBar* header_bar = GTK_HEADER_BAR(gtk_header_bar_new());
     gtk_widget_show(GTK_WIDGET(header_bar));
-    gtk_header_bar_set_title(header_bar, "LifeRhythm");
+    gtk_header_bar_set_title(header_bar, "DailyRhythm");
     gtk_header_bar_set_show_close_button(header_bar, TRUE);
     gtk_window_set_titlebar(window, GTK_WIDGET(header_bar));
   } else {
-    gtk_window_set_title(window, "LifeRhythm");
+    gtk_window_set_title(window, "DailyRhythm");
   }
 
   gtk_window_set_default_size(window, 1280, 720);
