+++ pubspec.yaml
@@ -1,5 +1,5 @@
-name: life_rhythm
-description: "LifeRhythm - Daily stats tracking app with monochrome rhythm aesthetic"
+name: daily_rhythm
+description: "DailyRhythm - Daily stats tracking app with monochrome rhythm aesthetic"
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
