+++ windows/runner/Runner.rc
@@ -89,13 +89,13 @@ BEGIN
     BEGIN
         BLOCK "040904e4"
         BEGIN
-            VALUE "CompanyName", "com.liferhythm" "\0"
-            VALUE "FileDescription", "LifeRhythm" "\0"
+            VALUE "CompanyName", "com.dailyrhythm" "\0"
+            VALUE "FileDescription", "DailyRhythm" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "LifeRhythm" "\0"
-            VALUE "LegalCopyright", "Copyright (C) 2025 com.liferhythm. All rights reserved." "\0"
-            VALUE "OriginalFilename", "LifeRhythm.exe" "\0"
-            VALUE "ProductName", "LifeRhythm" "\0"
+            VALUE "InternalName", "DailyRhythm" "\0"
+            VALUE "LegalCopyright", "Copyright (C) 2025 com.dailyrhythm. All rights reserved." "\0"
+            VALUE "OriginalFilename", "DailyRhythm.exe" "\0"
+            VALUE "ProductName", "DailyRhythm" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
